+++ src/MasterData.Core/DataManager/Exportation/Abstractions/IExportFormat.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using JJMasterData.Core.DataManager.Exportation;
 
 namespace JJMasterData.Core.DataManager.Exportation.Abstractions;
