+++ src/MasterData.Core/DataManager/Exportation/ExportFormatOptionKind.cs
@@ -1,4 +1,4 @@
-namespace JJMasterData.Core.DataManager.Exportation.Abstractions;
+namespace JJMasterData.Core.DataManager.Exportation;
 
 public enum ExportFormatOptionKind
 {
