+++ src/MasterData.Core/DataManager/Exportation/FormatOptionsBinder.cs
@@ -4,7 +4,6 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
-using JJMasterData.Core.DataManager.Exportation.Abstractions;
 
 namespace JJMasterData.Core.DataManager.Exportation;
