+++ src/MasterData.Core/DataManager/Exportation/Formats/DelimitedTextWriter.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using CsvHelper;
 using CsvHelper.Configuration;
 using JJMasterData.Core.DataDictionary.Models;
-using JJMasterData.Core.DataManager.Exportation.Abstractions;
+using JJMasterData.Core.DataManager.Exportation;
 using JJMasterData.Core.DataManager.Models;
 using JJMasterData.Core.DataManager.Services;
