+++ src/MasterData.Core/DataManager/Exportation/Formats/ExcelXlsxExportFormat.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using JJMasterData.Core.DataManager.Exportation;
 using JJMasterData.Core.DataManager.Exportation.Abstractions;
 using MiniExcelLibs;
 using MiniExcelLibs.OpenXml;
