+++ src/MasterData.Core/UI/Components/Exportation/DataExportationFactory.cs
@@ -6,7 +6,6 @@ using JJMasterData.Core.DataDictionary.Models;
 using JJMasterData.Core.DataDictionary.Repository.Abstractions;
 using JJMasterData.Core.DataManager;
 using JJMasterData.Core.DataManager.Exportation;
-using JJMasterData.Core.DataManager.Exportation.Abstractions;
 using JJMasterData.Core.DataManager.Exportation.Background;
 using JJMasterData.Core.DataManager.Expressions;
 using Microsoft.Extensions.Localization;
