+++ src/MasterData.Core/UI/Components/Exportation/DataExportationSettings.cs
@@ -13,7 +13,6 @@ using JJConsulting.Html.Extensions;
 using JJConsulting.MasterData.Storage.Abstractions;
 using JJMasterData.Core.DataManager.Exportation;
 using JJMasterData.Core.DataManager.Exportation.Configuration;
-using JJMasterData.Core.DataManager.Exportation.Abstractions;
 using Microsoft.Extensions.Localization;
 
 namespace JJMasterData.Core.UI.Components;
