+++ src/Plugins/MasterData.Pdf/MasterDataServiceBuilderExtensions.cs
@@ -6,6 +6,6 @@ public static class MasterDataServiceBuilderExtensions
 {
     public static MasterDataServiceBuilder WithPdfExportation(this MasterDataServiceBuilder builder)
     {
-        return builder.AddExportFormat<PdfWriter, PdfExportOptions>();
+        return builder.AddExportFormat<PdfExportFormat, PdfExportOptions>();
     }
 }
